requirements.txt
matplotlib==3.9.3
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
